Move moon button events to server id folder
